+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 - `Faker.Gov.It` [[@neslinesli93](https://github.com/neslinesli93)]
 - `Faker.Vehicle.model/1` [[@daytonn][https://github/com/daytonn]]
 - `Faker.Person`[[@anthonator](https://github.com/anthonator)]
+- `Faker.String.alphanumeric/0` and `Faker.String.alphanumeric/1`[[@brunohkbx](https://github.com/brunohkbx)]
 
 ### Changed
 - Fix `Faker.Code.Iban.iban` and `Faker.Gov.It.fiscal_id` doctests [[@vbrazo](https://github.com/vbrazo)]